functions-python/requirements.txt
firebase-functions
rembg
Pillow
numpy
onnxruntime
flask
cors
